+++ Basil/include/Renderer/Framebuffer.h
@@ -18,6 +18,7 @@ namespace Basil
 	class Framebuffer
 	{
 		public:
+			virtual ~Framebuffer() {}
 			virtual void bind() = 0;
 			virtual void unbind() = 0;
 			virtual void resize(uint32_t width, uint32_t height) = 0;